image type is a single asterisk; projects can be managed within the maker_mike project; moved home to the maker_mike project as well; the 2 navigations show UI-consistent toggle buttons

+++ src/maker_mike.project.d.php
@@ -4,6 +4,8 @@
 		delete database;
 	*/
 		$project = json_decode($row['config'])->_projectName;
-		exec('rm -r ../projects/'.$project);
-		exec('mysql -h '.$db_host.' -P '.$db_port.' -u '.$db_user.' --password='.$db_pass.' -e "DROP DATABASE IF EXISTS '.$project.';"');
+		if($project != 'maker_mike'){
+			exec('rm -r ../projects/'.$project);
+			exec('mysql -h '.$db_host.' -P '.$db_port.' -u '.$db_user.' --password='.$db_pass.' -e "DROP DATABASE IF EXISTS '.$project.';"');
+		}
 ?>
